requirements.txt
apify-client==1.8.1
python-dotenv==1.0.1
chromadb==0.6.3
google-generativeai==0.8.4
streamlit==1.41.1